VEGA-676: Доработать уведомления в vega-auth (#90)
* chore(global): убрать standalone сборку

Jira issue: VEGA-676

* feat(notifications): заменить нотификации на vega-shell

Jira issue: VEGA-676
