+++ package.json
@@ -18,9 +18,8 @@
     }
   },
   "scripts": {
-    "build": "cross-env NODE_ENV=production webpack --mode production --progress --config ./webpack.forShellConfig.js",
-    "dev": "cross-env NODE_ENV=development webpack-dev-server --open --mode development",
-    "dev:shell": "cross-env NODE_ENV=development webpack-dev-server --port 9002 --mode development  --config ./webpack.forShellConfig.js",
+    "build": "cross-env NODE_ENV=production webpack --mode production --progress --config ./webpack.config.js",
+    "dev": "cross-env NODE_ENV=development webpack-dev-server --port 9002 --mode development  --config ./webpack.config.js",
     "test": "jest",
     "test:watch": "jest --watch",
     "format": "prettier --write {src,types,scripts}/**/*.{json,{j,t}s{,x},css,}",
