+++ src/App/App.test.tsx
@@ -1,13 +1,26 @@
 import React from 'react';
 import { render } from '@testing-library/react';
 
+import { Unsubscribe } from '../../types/notifications';
+
 import { App } from './App';
+
+const noop = (): void => {};
+
+const identity = {
+  getToken: noop,
+};
+
+const notifications = {
+  add: noop,
+  remove: noop,
+  subscribe: (): Unsubscribe => {
+    return (): void => {};
+  },
+};
 
 describe('App', () => {
   test('корректно рендерится', () => {
-    const identity = {
-      getToken: () => 'token',
-    };
-    render(<App identity={identity} />);
+    render(<App identity={identity} notifications={notifications} />);
   });
 });
