+++ src/platform/index.ts
@@ -1,2 +1,1 @@
-export * from './snackbar';
 export * from './app-context';
