--- src/platform/snackbar/Snackbar.css
@@ -1,15 +1,0 @@
-.vega-auth-snackbar {
-  width: 400px;
-}
-
-.vega-auth-snackbar-portal {
-  position: fixed;
-  z-index: 100;
-
-  bottom: 0;
-
-  display: flex;
-  justify-content: center;
-
-  width: 100%;
-}
