--- src/platform/snackbar/index.ts
@@ -1,1 +1,0 @@
-export { SnackbarProvider, useSnackbar } from './Snackbar';
